Added DB components
The database schema as well as some core table data to get people
started.

NOTE: The VS project has the database set to NOT COPY in order to get it
to build without the database file in the project (excluded because it
might contain login info). In your own project, you might want to turn
it back on, but it will possibly break other people's builds.. they need
to be aware of this.