Finishing touches (mostly) on new format
The new Image based first tab is just about completed. The top 5 matches
are filled in when you expand, I just need to verify the logic/code for
scrapes works (it should) and pushes out expire matches from the list.

Expanding a group fills the matches on the first run.. scraping puts the
new match at the front of the list and refreshes the panels. Also set
the color to firebrick on expanded groups so they are easier to see, and
clicking keeps focus on the tab page (mouse wheel scrolling behaves
better).

todo: try to find a "click" on the interior group boxes to make it even
easier to mouse wheel..